Reorg routes other than dataUploads
