+++ backend/http-requests/repo.http
@@ -1,10 +1,11 @@
 
 ### /datauploads - default flow
 POST {{base-url}}/repos
+Authorization: Bearer {{jwt}}
 Content-Type: application/json
 
 {
-  "data_upload_id": "xyz",
+  "data_upload_id": "5e7a45e510abf3bbf5068f6f",
   "name": "Dataset 1 - Upload 1 repo"
 }
