+++ backend/package.json
@@ -28,6 +28,7 @@
     "nodemailer": "^6.3.0",
     "npmlog": "^4.1.2",
     "passport": "^0.4.1",
+    "passport-jwt": "^4.0.0",
     "passport-openidconnect": "0.0.2",
     "tableschema": "^1.11.2"
   },
